begin project simplification
