+++ .gitignore
@@ -14,3 +14,25 @@ neutralino.js
 
 # Ignore temp files, these get automatically generated
 /.tmp
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
