+++ neutralino.config.json
@@ -1,14 +1,11 @@
 {
-  "$schema": "https://raw.githubusercontent.com/neutralinojs/neutralinojs/main/schemas/neutralino.config.schema.json",
   "applicationId": "js.neutralino.sample",
   "version": "1.0.0",
   "defaultMode": "window",
-  "port": 0,
-  "documentRoot": "/react-src/build/",
+  "documentRoot": "/build/",
   "url": "/",
   "enableServer": true,
   "enableNativeAPI": true,
-  "tokenSecurity": "one-time",
   "logging": {
     "enabled": true,
     "writeToLogFile": true
@@ -16,6 +13,7 @@
   "nativeAllowList": [
     "app.*",
     "os.*",
+    "events.*",
     "debug.log"
   ],
   "globalVariables": {
@@ -39,44 +37,28 @@
       "minHeight": 200,
       "fullScreen": false,
       "alwaysOnTop": false,
-      "icon": "/resources/icons/appIcon.png",
+      "icon": "/public/logo192.png",
       "enableInspector": true,
       "borderless": false,
       "maximize": false,
       "hidden": false,
-      "resizable": true,
-      "exitProcessOnClose": true
-    },
-    "browser": {
-      "globalVariables": {
-        "TEST": "Test value browser"
-      },
-      "nativeBlockList": [
-        "filesystem.*"
-      ]
-    },
-    "cloud": {
-      "url": "/resources/#cloud",
-      "nativeAllowList": [
-        "app.*"
-      ]
-    },
-    "chrome": {
-      "width": 800,
-      "height": 500,
-      "args": "--user-agent=\"Neutralinojs chrome mode\"",
-      "nativeBlockList": [
-        "filesystem.*",
-        "os.*"
-      ]
+      "resizable": true
     }
   },
   "cli": {
     "binaryName": "raven",
-    "resourcesPath": "/react-src/build",
+    "resourcesPath": "/build",
     "extensionsPath": "/extensions/",
-    "clientLibrary": "/www/neutralino.js",
+    "clientLibrary": "/public/neutralino.js",
     "binaryVersion": "4.14.1",
-    "clientVersion": "3.12.0"
+    "clientVersion": "3.12.0",
+    "frontendLibrary": {
+      "patchFile": "/public/index.html",
+      "devUrl": "http://localhost:3000",
+      "projectPath": "./",
+      "initCommand": "npm install",
+      "devCommand": "npm start",
+      "buildCommand": "npm run build"
+  }
   }
 }
