+++ package.json
@@ -1,11 +1,14 @@
 {
-  "name": "react-src",
+  "name": "raven",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@neutralinojs/lib": "^3.13.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "cross-env": "^7.0.3",
+    "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
@@ -34,5 +37,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "java-parser": "^2.2.0"
   }
 }
