+++ public/index.html
@@ -9,6 +9,8 @@
       name="description"
       content="Web site created using create-react-app"
     />
+    <script src="http://localhost:50388/__neutralino_globals.js"></script>
+    <script src="http://localhost:50388/neutralino.js"></script>
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
